+++ jaxrs-base-server/src/test/java/io/telicent/smart/cache/server/jaxrs/applications/TestServer.java
@@ -35,14 +35,19 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import org.apache.commons.lang3.Strings;
 import org.apache.jena.riot.web.HttpNames;
+import org.glassfish.jersey.http.HttpHeaders;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.ConnectException;
 import java.util.Map;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
 import java.util.concurrent.*;
 import java.util.function.Function;
 
@@ -53,6 +58,7 @@ public class TestServer extends AbstractAppEntrypoint {
     private static final RandomPortProvider PORT = new RandomPortProvider(1366);
 
     private final Client client = ClientBuilder.newClient().register(ProblemCustomReaderWriter.class);
+    private final HttpClient jdkClient = HttpClient.newHttpClient();
 
     @BeforeMethod
     public void setup() {
@@ -573,7 +579,7 @@ public class TestServer extends AbstractAppEntrypoint {
     }
 
     @Test
-    public void givenServer_whenPostingEmptyBody_then400BadRequest() throws IOException, InterruptedException {
+    public void givenServer_whenPostingEmptyBodyToResourceThatRequiresRequestBody_then400BadRequest() throws IOException {
         // Given
         ServerBuilder builder = buildServer();
         try (Server server = builder.build()) {
@@ -593,7 +599,8 @@ public class TestServer extends AbstractAppEntrypoint {
     }
 
     @Test
-    public void givenServer_whenPostingWrongFormatBody_then415UnsupportedMediaType() throws IOException, InterruptedException {
+    public void givenServer_whenPostingWrongFormatBody_then415UnsupportedMediaType() throws IOException,
+            InterruptedException {
         // Given
         ServerBuilder builder = buildServer();
         try (Server server = builder.build()) {
@@ -614,7 +621,7 @@ public class TestServer extends AbstractAppEntrypoint {
     }
 
     @Test
-    public void givenServer_whenPostingEmptyBodyToResourceThatDoesNotRequireABody_thenOK() throws IOException, InterruptedException {
+    public void givenServer_whenPostingEmptyBodyToResourceThatDoesNotRequireABody_thenOK() throws IOException {
         // Given
         ServerBuilder builder = buildServer();
         try (Server server = builder.build()) {
@@ -630,6 +637,51 @@ public class TestServer extends AbstractAppEntrypoint {
                 MockData data = response.readEntity(MockData.class);
                 Assert.assertEquals(data.value(), "test");
             }
+        }
+    }
+
+    @Test
+    public void givenServer_whenPostingWithNoContentType_then400BadRequest() throws IOException,
+            InterruptedException {
+        // Given
+        ServerBuilder builder = buildServer();
+        try (Server server = builder.build()) {
+            server.start();
+            // When
+            WebTarget target = forServer(server, "/params/everything/test");
+            HttpRequest request =
+                    HttpRequest.newBuilder()
+                               .uri(target.getUri())
+                               .method("POST", HttpRequest.BodyPublishers.noBody())
+                               .header(HttpHeaders.CONTENT_TYPE, "")
+                               .build();
+            HttpResponse<InputStream> response = jdkClient.send(request, HttpResponse.BodyHandlers.ofInputStream());
+
+            // Then
+            Assert.assertEquals(response.statusCode(), Response.Status.BAD_REQUEST.getStatusCode());
+        }
+    }
+
+    @Test
+    public void givenServer_whenPostingWithMalformedContentType_then400BadRequest() throws IOException,
+            InterruptedException {
+        // Given
+        ServerBuilder builder = buildServer();
+        try (Server server = builder.build()) {
+            server.start();
+
+            // When
+            WebTarget target = forServer(server, "/data/test").queryParam("value", "test");
+            HttpRequest request =
+                    HttpRequest.newBuilder()
+                               .uri(target.getUri())
+                               .method("POST", HttpRequest.BodyPublishers.noBody())
+                               .header(HttpHeaders.CONTENT_TYPE, "undefined")
+                               .build();
+            HttpResponse<InputStream> response = this.jdkClient.send(request, HttpResponse.BodyHandlers.ofInputStream());
+
+            // Then
+            Assert.assertEquals(response.statusCode(), Response.Status.BAD_REQUEST.getStatusCode());
         }
     }
 
@@ -761,14 +813,15 @@ public class TestServer extends AbstractAppEntrypoint {
     }
 
     public static void verifyProblemContent(Problem problem, String expectedTitle, String expectedType,
-                                             String expectedDetail) {
+                                            String expectedDetail) {
         Assert.assertEquals(problem.getTitle(), expectedTitle);
         Assert.assertEquals(problem.getType(), expectedType);
         Assert.assertTrue(Strings.CI.contains(problem.getDetail(), expectedDetail));
     }
 
     @Test
-    public void givenServer_whenGeneratingAProblemResponseWithNoAcceptHeader_thenProblemIsReturnedInProblemContentType() throws IOException {
+    public void givenServer_whenGeneratingAProblemResponseWithNoAcceptHeader_thenProblemIsReturnedInProblemContentType() throws
+            IOException {
         // Given
         ServerBuilder builder = buildServer();
         try (Server server = builder.build()) {
@@ -789,7 +842,8 @@ public class TestServer extends AbstractAppEntrypoint {
     }
 
     @Test
-    public void givenServer_whenGeneratingAProblemResponseWithWildcardAcceptHeader_thenProblemIsReturnedInProblemContentType() throws IOException {
+    public void givenServer_whenGeneratingAProblemResponseWithWildcardAcceptHeader_thenProblemIsReturnedInProblemContentType() throws
+            IOException {
         // Given
         ServerBuilder builder = buildServer();
         try (Server server = builder.build()) {
@@ -889,7 +943,7 @@ public class TestServer extends AbstractAppEntrypoint {
             server.start();
 
             // When
-            WebTarget target = forServer(server,"/problems/bad-annotations");
+            WebTarget target = forServer(server, "/problems/bad-annotations");
             Invocation.Builder invocation = target.request(MediaType.APPLICATION_JSON);
 
             // Then
@@ -1004,7 +1058,7 @@ public class TestServer extends AbstractAppEntrypoint {
     // Can be useful to run the server in blocking mode for debugging, but commented out by default as otherwise the IDE
     // won't try to run the class as a Test class by default
 
-    /*
+/*
     public static void main(String[] args) {
         TestServer test = new TestServer();
         ServerBuilder builder = test.buildServer().withVersionInfo("observability-core", "jaxrs-base-server");
@@ -1014,5 +1068,7 @@ public class TestServer extends AbstractAppEntrypoint {
         } catch (IOException | InterruptedException e) {
             e.printStackTrace();
         }
-    }*/
+    }
+
+ */
 }